+++ include/unity/scopes/internal/ScopeObject.h
@@ -77,8 +77,8 @@ public:
                                      MWReplyProxy const& reply,
                                      InvokeInfo const& info) override;
 
-    virtual ChildScopeList child_scopes_ordered() const override;
-    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) override;
+    virtual ChildScopeList child_scopes() const override;
+    virtual bool set_child_scopes(ChildScopeList const& child_scopes) override;
 
     virtual bool debug_mode() const override;