+++ include/unity/scopes/internal/zmq_middleware/ZmqScope.h
@@ -63,8 +63,8 @@ public:
                                    VariantMap const& hints,
                                    MWReplyProxy const& reply) override;
 
-    virtual ChildScopeList child_scopes_ordered() override;
-    virtual bool set_child_scopes_ordered(ChildScopeList const& child_scopes_ordered) override;
+    virtual ChildScopeList child_scopes() override;
+    virtual bool set_child_scopes(ChildScopeList const& child_scopes) override;
 
     virtual bool debug_mode() override;